+++ src/components/theme/Header/Header.test.jsx
@@ -7,7 +7,7 @@ jest.mock('../SearchWidget/SearchWidget', () => jest.fn(() => <div />));
 jest.mock('../Anontools/Anontools', () => jest.fn(() => <div />));
 
 test('renders a header component', () => {
-  const component = renderer.create(<Header />);
+  const component = renderer.create(<Header pathname="/blog" />);
   const json = component.toJSON();
   expect(json).toMatchSnapshot();
 });