+++ src/middleware/api.js
@@ -23,14 +23,15 @@ export default api => ({ dispatch, getState }) => next => action => {
   next({ ...rest, type: `${type}_PENDING` });
 
   const actionPromise = promise(api);
-  actionPromise
-    .then(
-      result => next({ ...rest, result, type: `${type}_SUCCESS` }),
-      error => next({ ...rest, error, type: `${type}_ERROR` }),
-    )
+  actionPromise.then(
+    result => next({ ...rest, result, type: `${type}_SUCCESS` }),
+    error => next({ ...rest, error, type: `${type}_ERROR` }),
+  );
+  /*
     .catch(error => {
       next({ ...rest, error, type: `${type}_ERROR` });
     });
+    */
 
   return actionPromise;
 };