+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+### Added
+- Delete in folder contents @robgietema
 
 ## 0.1.0 (2017-04-20)
 ### Added